app/backend/app/intro/serializers/post.py
from rest_framework import serializers

from app.intro.models import Post


class PostModelSerializer(serializers.ModelSerializer):
    class Meta:
        model = Post
        fields = '__all__'
        