requirements.txt
Django==3.2.14
djangorestframework==3.13.1
drf-yasg==1.20.0
djangorestframework-simplejwt==5.2.0
django-environ==0.9.0
django-cors-headers==3.13.0
psycopg2-binary==2.9.3
djangorestframework-filters==1.0.0.dev2
Pillow==9.2.0
django-anymail==7.1.0
boto3==1.24.40
django-storages==1.12.3
